refactor: Rename `…b/plugins/aws/package/lib/addExportNameForOutputs.js`
