+++ lib/plugins/aws/package/index.js
@@ -8,7 +8,7 @@ const generateCoreTemplate = require('./lib/generateCoreTemplate');
 const saveServiceState = require('./lib/saveServiceState');
 const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
 const mergeIamTemplates = require('./lib/mergeIamTemplates');
-const addExportNameForOutputs = require('./lib/addExportNameForOutputs');
+const addExportNameForOutputs = require('./lib/add-export-name-for-outputs');
 const { log, style, progress } = require('@serverless/utils/log');
 
 const mainProgress = progress.get('main');
