+++ lib/plugins/aws/package/lib/add-export-name-for-outputs.js
