+++ test/unit/lib/plugins/aws/package/lib/add-export-name-for-outputs.test.js
